railway-world-class/backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
bcrypt==4.1.2
PyJWT==2.9.0
celery==5.4.0
redis==5.0.7
httpx==0.27.2
reportlab==4.2.2
python-multipart==0.0.9
playwright==1.49.0
